requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
pinecone-client==3.0.0
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncpg==0.29.0
redis==5.0.1
numpy==1.26.0
scikit-learn==1.3.2
aiohttp==3.9.0
python-multipart==0.0.6
requests==2.31.0
